blank_project/requirements.txt
Django==1.9.2
djangorestframework==3.3.2
Pillow==3.1.1
Pygments==2.1
wheel==0.24.0
